% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_naei.R
\name{read_naei}
\alias{read_naei}
\title{Read NAEI Data}
\usage{
read_naei()
}
\value{
The NAEI data as a data frame.
}
\description{
This function reads NAEI (National Atmospheric Emissions Inventory) data
}
